Allow a page to act as the navigation root
A new property called "Page is navigation root" is added to the settings
tab. When checked the page becomes the effective navigation root shows
as the start page of the bread crumb and also affects the left nav.
Useful for defining sets of pages as microsites. The breadcrumb can also
be clicked in on page edit to set and unset the flag.